Frontend for currency converter

+++ CurrencyConverter.Api/Api/Data/AppSettings.cs
@@ -8,7 +8,7 @@ namespace OlegChibikov.OctetInterview.CurrencyConverter.Api.Data
 
         public string? CurrencyConverterApiKey { get; set; }
 
-        public double MarkupPercentage { get; set; } = 0.1;
+        public decimal MarkupPercentage { get; set; } = 0.1M;
 
         public TimeSpan CacheDuration { get; set; } = TimeSpan.FromHours(1);
     }
