+++ CurrencyConverter.Api/Contracts/IConversionRateRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IConversionRateRepository
     {
-        void SaveRate(string sourceCurrencyCode, string targetCurrencyCode, double rate);
+        void SaveRate(string sourceCurrencyCode, string targetCurrencyCode, decimal rate);
 
         ConversionRate? GetRate(string sourceCurrencyCode, string targetCurrencyCode);
     }
